distance done, but error when i want to test ...

+++ src/Testing.hh
@@ -7,7 +7,7 @@
 
 namespace Testing
 {
-    int distance(Sign s1, Sign s2);
+	double distance(Sign s1, Sign s2);
 
     void test();
 }
